HeatMapChart: migrated enzyme testst to react tests
